Start adding the Fortran interface to Verificarlo
vfc_probes has been coverted to a shared object library, and the source
code of the Fortran interface has also been added.

Now, it needs to be added to the install process, only when the
--with-flang is specified.
