+++ src/common/Makefile.am
@@ -15,4 +15,11 @@ libvfc_hashmap_la_SOURCES = \
 	vfc_hashmap.h \
 	vfc_hashmap.c
 
+
+lib_LTLIBRARIES = libvfc_probes.la
+libvfc_probes_la_SOURCES = vfc_hashmap.c vfc_probes.c
+if WALL_CFLAGS
+libvfc_probes_la_CFLAGS = -Wall -Wextra
+endif
+
 library_includedir =$(includedir)/
