+++ tests/test_fortran_vfc_probes/clean.sh
@@ -1,0 +1,1 @@
+verificarlo-f vfc_probes_test.f90 vfc_probes.so -o vfc_probes_test
