+++ tests/test_fortran_vfc_probes/test.sh
@@ -1,0 +1,2 @@
+flang -c vfc_probes_test.f90
+flang vfc_probes_test.o libvfc_probes.so -o fortran_test
