--- tests/test_static_mode/clean.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-rm -Rf *~ test.log
