--- tests/test_static_mode/test.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-set -e
-verificarlo-c -static -O0 test.c -o test
-VFC_BACKENDS="libinterflop_ieee.so --debug" ./test
